CRIUと計測方法を合わせる

+++ core/iwasm/migration/wasm_restore.h
@@ -4,7 +4,7 @@
 #include "../common/wasm_exec_env.h"
 #include "../interpreter/wasm_interp.h"
 
-double get_restore_framestack_time();
+long int get_restore_framestack_time();
 
 void set_restore_flag(bool f);
 bool get_restore_flag();
